Prepare for more structured routing
We're about to make routes return more structured information if they don't match, to help the router know whether to check the next route,
or give up and return an error response.

At the moment routes return a message with the json path to the bit of the request that didn't match. This is baked into Result.

We need to support lenses returning a Result with different types for errors (i.e. be a proper port of Result4k)

* Rename existing Problems stuff to JsonPath*
* Continue to use json path for route failures for now, although we're about to stop doing that